chore: rename lib/ioredis.js to lib/redis.js
