+++ lib/redis.js
@@ -1,18 +1,12 @@
-var EventEmitter = require('events')
-var log = require('@dadi/logger')
-var redis = require('redis')
-var redisRStream = require('redis-rstream')
-var redisWStream = require('redis-wstream')
-var Stream = require('stream')
-var streamifier = require('streamifier')
-var util = require('util')
+const util = require('util')
+const EventEmitter = require('events')
+const redis = require('ioredis')
+const redisRStream = require('redis-rstream')
+const redisWStream = require('redis-wstream')
+const Stream = require('stream')
+const streamifier = require('streamifier')
+const log = require('@dadi/logger')
 
-/**
- * Creates a new RedisCache instance
- * @constructor
- * @classdesc This is a description of the RedisCache.
- * @param {object} options - the options used to instantiate a RedisCache handler
- */
 function RedisCache (options) {
   this.options = options
   this.ttl = options.ttl || 3600
@@ -23,14 +17,60 @@ function RedisCache (options) {
 
 util.inherits(RedisCache, EventEmitter)
 
-/**
- * @param {String} key - the key to retrieve from cache
- * @returns {Promise.<Stream, Error>} A promise that returns a Stream if the key exists,
- *     or an Error if it does not exist
- */
-RedisCache.prototype.get = function (key) {
-  return new Promise((resolve, reject) => {
+RedisCache.prototype.initialise = function initialise(options) {
+  const self = this
+  this.timesConnected = 0
 
+  function retryStrategy(attempts) {
+    const baseRetryTime = 1024
+    const maxRetryTime = 4096
+    const maxConnectedTimes = 3
+    const currentRetryTime = baseRetryTime * attempts
+
+    if (currentRetryTime > maxRetryTime) {
+      self.onFailure()
+      return new Error('Exceeded max retry time')
+    }
+
+    if (self.timesConnected > maxConnectedTimes) {
+      self.onFailure()
+      return new Error('Exceeded max times connected; Redis appears unstable')
+    }
+
+    return currentRetryTime
+  }
+
+  const redisClient = redis.createClient(options.port, options.host, {
+    detect_buffers: true,
+    retry_strategy: retryStrategy
+  })
+
+  // error: doesn't get fired on dc errors
+  redisClient.on('error', (err) => {
+    log.error(err)
+  })
+
+  // end: should fire only on graceful dc
+  redisClient.on('end', () => {
+    this.emit('end')
+  })
+
+  // ready: when we are connected
+  redisClient.on('ready', () => {
+    this.timesConnected++
+    this.emit('ready')
+  })
+
+  // reconnecting: every attempt
+  redisClient.on('reconnecting', (attempt) => {
+    this.emit('reconnecting', attempt)
+  })
+
+  return redisClient
+}
+
+RedisCache.prototype.get = function get(key) {
+  return new Promise((resolve, reject) => {
     if (!this.redisClient.connected) {
       this.emit('fail')
       return reject(new Error('The specified key does not exist'))
@@ -38,7 +78,7 @@ RedisCache.prototype.get = function (key) {
 
     this.redisClient.exists(key, (err, exists) => {
       if (exists > 0) {
-        var stream = redisRStream(this.redisClient, key)
+        const stream = redisRStream(this.redisClient, key)
         return resolve(stream)
       } else {
         return reject(new Error('The specified key does not exist'))
@@ -47,17 +87,13 @@ RedisCache.prototype.get = function (key) {
   })
 }
 
-/**
- * @param {String} key - the key to store the data against
- * @param {String|Buffer|Stream} data - the data to cache, as a String, Buffer or Stream
- * @returns {Promise.<String, Error>} A promise that returns an empty String if successful, otherwise an Error
- */
-RedisCache.prototype.set = function (key, data) {
-
-  if (!this.redisClient.connected) this.emit('fail', 'set', key, data)
+RedisCache.prototype.set = function set(key, data) {
+  if (!this.redisClient.connected) {
+    this.emit('fail', 'set', key, data)
+  }
 
   return new Promise((resolve, reject) => {
-    var redisWriteStream = redisWStream(this.redisClient, key)
+    const redisWriteStream = redisWStream(this.redisClient, key)
 
     redisWriteStream.on('finish', () => {
       if (this.ttl) {
@@ -82,56 +118,8 @@ RedisCache.prototype.set = function (key, data) {
   })
 }
 
-/**
- * Initialises a RedisClient using the main configuration settings
- * @returns {RedisClient}
- */
-RedisCache.prototype.initialise = function (options) {
-  var self = this
-
-  function retryStrategy (options) {
-    var baseRetryTime = 1024
-    var maxRetryTime = 4096
-    var maxConnectedTimes = 3
-
-    var currentRetryTime = baseRetryTime * options.attempt
-
-    if (currentRetryTime > maxRetryTime) {
-      self.onFailure()
-      return new Error('Exceeded max retry time')
-    }
-
-    if (options.times_connected > maxConnectedTimes) {
-      self.onFailure()
-      return new Error('Exceeded max times connected; Redis appears unstable')
-    }
-
-    return currentRetryTime
-  }
-
-  var redisClient = redis.createClient(options.port, options.host, {detect_buffers: true, retry_strategy: retryStrategy})
-
-  redisClient.on('error', function (err) { // Doesn't get fired on dc errors
-    log.error(err)
-  })
-
-  redisClient.on('end', () => { // should fire only on graceful dc
-    this.emit('end')
-  })
-
-  redisClient.on('ready', () => { // when we are connected
-    this.emit('ready')
-  })
-
-  redisClient.on('reconnecting', (attempt) => { // every attempt
-    this.emit('reconnecting', attempt)
-  })
-
-  return redisClient
-}
-
-RedisCache.prototype.onFailure = function () {
-  var redisRetryTime = 1000 * 60 * 1
+RedisCache.prototype.onFailure = function onFailure() {
+  const redisRetryTime = 1000 * 60 * 1
 
   this.emit('fail')
